RETOPARCIAL/requirements.txt
firebase-admin>=6.5.0,<7
python-dotenv>=1.0.1,<2

